rebase: extract rebaseset and destination computation in a function
The whole rebase function is gargantuan and this computation is almost 100 lines
long.  We extract it in a dedicated function as it is independent from the rest
of the rebase code.

Having it in its own function will make it easier to rework the default
destination logic.

--HG--
extra : rebase_source : 9fe8e0e9acbb697c163caf4a629249fc412a2985