graphmod: augment the graph to include more information about the edges
The walker knows when an edge leads to a direct parent, a grandparent (skipping
revisions not part of the revset) and parents that are missing altogether
(neither it nor a grandparent is in the revset). Add this information to the
parents sequence yielded.

--HG--
extra : rebase_source : c487bf64cb0f30f3eb5de33dc126eb0ea9f0f49e