chgserver: implement validate command
validate will load the repo config and check if the server has up-to-date
config to continue serve the client. In case it does not, the server will
send instructions to the client about what to do next, including to retry
with a different address or to unlink an outdated socket file to stop an
old server.

--HG--
extra : amend_source : a9dc4c79a81ad72ded9b41dd40059dd2a114d94f