changelog: parse description last
Before, we first searched for the double newline as the first step in
the parse then moved to the front of the string and worked our way
to the back again. This made sense when we were splitting the raw
text on the double newline. But in our new newline scanning based
approach, this feels awkward.

This patch updates the parsing logic to parse the text linearly and
deal with the description field last.

Because we're avoiding an extra string scan, revsets appear to
demonstrate a very slight performance win. But the percentage
change is marginal, so the numbers aren't worth reporting.

--HG--
extra : rebase_source : dcf114b97ca637f5ac95673c6829da1e93070e37