extensions: factor import error reporting out
To clarify third party extensions lookup, we are about to add a third place
where extensions are searched for. So we factor the error reporting logic out to
be able to easily reuse it in the next patch.

--HG--
extra : rebase_source : aa21d443e12ce92f18a42ad8230d365aa839fb66