hg: perform update after pulling during clone with share (issue5103)
This is a graft of 60adda1a0188 and a4692267bc2d from the default
branch. Combined, they address a bug with pooled shared storage where
an update may not update to the most recent revision when performing
a `hg clone`.

The patches should have been written against the stable branch in
the beginning. I screwed up.

--HG--
branch : stable