run-tests: make _processoutput picky about optional globs
1ad0ddf8cccc enabled lines that were not matched to be found later in cases of jitter.

Unfortunately, in this model an optional line would always jitter to the end when
it is not present. That is not ideal.

It would be possible to do better, by queuing all writes until the end in case
an optional line jitters, but for now, it is simpler to assume optional lines
have a fixed place in the stream.

--HG--
extra : rebase_source : 866c2ff10b0345e59e3f137d5e79991b1befc678
extra : histedit_source : 9f2516da2a99ffcbf1af562338268c5d8ccf9f9f