repoview: ignore unwritable hidden cache
The atomictemp.close() file attempts to do a rename, which can fail.
Moving the close inside the exception handler fixes it.

This doesn't fit well with the with: pattern, as it's the finalizer
that's failing.

--HG--
branch : stable
extra : rebase_source : 1ece0150b7e606db2415ec451adea68e8e7ad9ea