update: teach hg to override untracked dir with a tracked file on update
This is a fix to an old problem when Mercurial got confused by an
untracked folder with the same name as one of the files in a commit
hg was trying to update to. It is pretty safe to remove this folder if
it is empty. Backing up an empty folder seems to go against Mercurial's
"don't track dirs" philosophy.

--HG--
rename : tests/test-update-renames.t => tests/test-update-names.t
extra : amend_source : 66855c291f1a0e8a2c6fbe88aaeae8335861547a